weather-api/requirements.txt
aiomysql==0.2.0
fastapi==0.115.14
python-dotenv==1.1.1
uvicorn==0.34.3
httpx==0.27.0
